+++ Lem_in/libft/ft_atoi.c
@@ -3,20 +3,20 @@
 /*                                                              /             */
 /*   ft_atoi.c                                        .::    .:/ .      .::   */
 /*                                                 +:+:+   +:    +:  +:+:+    */
-/*   By: jde-la-m <jde-la-m@student.le-101.fr>      +:+   +:    +:    +:+     */
+/*   By: nsalle <nsalle@student.le-101.fr>          +:+   +:    +:    +:+     */
 /*                                                 #+#   #+    #+    #+#      */
 /*   Created: 2018/10/03 13:45:05 by jde-la-m     #+#   ##    ##    #+#       */
-/*   Updated: 2018/10/08 01:49:58 by jde-la-m    ###    #+. /#+    ###.fr     */
+/*   Updated: 2019/11/20 11:28:44 by nsalle      ###    #+. /#+    ###.fr     */
 /*                                                         /                  */
 /*                                                        /                   */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_atoi(const char *str)
+long long int	ft_atoi(const char *str)
 {
-	int res;
-	int sign;
+	long long int	res;
+	int				sign;
 
 	res = 0;
 	while (ft_isspace(*str))
@@ -25,5 +25,7 @@ int	ft_atoi(const char *str)
 	*str == '-' || *str == '+' ? str++ : str;
 	while (ft_isdigit(*str))
 		res = (res * 10) + (*str++ - '0');
+	// ft_putnbr(res);
+	// ft_putendl("");
 	return (res * sign);
 }
